build: eliminate all compiler warnings across every platform
The Linux/gcc build was already clean, but the clang-based targets
(macOS/iOS, Android, web) and the raylib dependency build emitted warnings.
Resolve every one so all CI compile logs are warning-free:

Our code:
- sound.c: cast RAND_MAX to float (clang -Wimplicit-const-int-float-conversion;
  2147483647 isn't representable in float).
- gfx_metal.mm: cast the '?' fallback glyph lookup to unsigned char
  (-Wchar-subscripts under Obj-C++).

Vendored single-header libs (compiled by our TUs):
- encode_h264.c: also suppress -Wpointer-to-int-cast (minih264, clang).
- encode_mux.c / recorder.c: also suppress -Wtypedef-redefinition (minimp4,
  clang-only under c99), guarded by __clang__ so GCC's -Wpragmas stays quiet.

NDK glue (compiled by our android target):
- Makefile: build native_app_glue.o with -Wno-unused-parameter (NDK source).

raylib's own build (its bundled stb_vorbis / jar_mod / __ANDROID_API__ redefine):
- build_raylib_*.sh: pass CUSTOM_CFLAGS=-w to raylib's make (appended after its
  -Wall, so it wins). -w only suppresses diagnostics, never changes codegen.

Verified locally with both gcc and a fresh clang 18: 0 warnings across all
openblocks C sources and the test suite; raylib relinks and `make test` passes.
The macOS/iOS/Android/web platform fixes are validated by CI.

+++ src/sound.c
@@ -62,7 +62,7 @@ static AudioHandle make_noise(float dur, float vol) {
     float hold = 0.0f;
     for (int i = 0; i < n; i++) {
         if (i % 8 == 0) { // sample-and-hold gives a grittier, lower-rate hiss
-            hold = ((float)rand() / RAND_MAX) * 2.0f - 1.0f;
+            hold = ((float)rand() / (float)RAND_MAX) * 2.0f - 1.0f;
         }
         float env = 1.0f - (float)i / n;
         buf[i] = (int16_t)(hold * vol * env * 32767.0f);
